+++ apps/ledger-live-mobile/src/screens/Manager/index.tsx
@@ -100,7 +100,7 @@ const ChooseDevice: React.FC<ChooseDeviceProps> = ({ isFocused }) => {
           <Trans i18nKey="manager.title" />
         </Text>
         {newDeviceSelectionFeatureFlag?.enabled ? (
-          <SelectDevice2 onSelect={onSelectDevice} />
+          <SelectDevice2 onSelect={onSelectDevice} stopBleScanning={!!device} />
         ) : (
           <>
             <SelectDevice